edited getCurrentTime method to return microseconds instead of seconds

+++ src/helpers/benchmark.h
@@ -14,6 +14,6 @@
   Methods
  **************************************/
 
-double getCurrentTime();
+unsigned long long getCurrentTime();
 
 #endif
